pyproject.toml
[project]
name = "cb-schedule"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.0",
    "img2table[paddle]>=1.4.2",
    "jinja2>=3.1.6",
    "opencv-python>=4.12.0.88",
    "paddleocr>=3.2.0",
    "paddlepaddle>=3.1.1",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "selectolax>=0.3.34",
]

[project.scripts]
ctc-schedule = "cb_schedule.services.ctc.parse_schedule_image:main"
cbl-schedule = "cb_schedule.services.cbl.scrape_schedule:main"
render-day = "cb_schedule.render_day:main"
publish = "cb_schedule.publish:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "ty>=0.0.1a20",
]

[tool.ruff]
line-length = 120

[tool.lint.ruff]
select = [
  "F",    # https://docs.astral.sh/docs/rules/#pyflakes-f
  "E",    # https://docs.astral.sh/docs/rules/#error-e
  "W",    # https://docs.astral.sh/docs/rules/#warning-w
  "B",    # https://docs.astral.sh/docs/rules/#flake8-bugbear-b
  "I",    # https://docs.astral.sh/docs/rules/#isort-i
  "NPY",  # https://docs.astral.sh/docs/rules/#numpy-specific-rules-npy
  "PD",   # https://docs.astral.sh/docs/rules/#pandas-vet-pd
  "RUF",  # https://docs.astral.sh/docs/rules/#ruff-specific-rules-ruf
  "T100",  #https://docs.astral.sh/ruff/rules/debugger/
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
